requirements.txt
# Backend
fastapi>=0.109.0
uvicorn[standard]>=0.22.0

# Multilingual embeddings and offline RAG
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.26.4
pandas>=2.1.0

# Online LLM APIs
openai>=1.7.3
groq>=0.9.0

# File uploads in FastAPI
python-multipart>=0.0.6

# Speech recognition (voice input)
SpeechRecognition>=3.9.0
pydub>=0.25.1

# Language detection
langdetect>=1.0.9

# Environment / config helpers
python-dotenv>=1.0.0

langchain
langchain-huggingface